Check if DB had the guild bot was deleted from.

+++ events/guildDelete.js
@@ -3,9 +3,9 @@ class GuildDelete extends require("../automation/eventClass"){
         super("guildDelete",false);
     }
     exec(guild,bot){
+        if(!bot.db.has(guild.id)) return;
         bot.db.delete(guild.id);
         console.log(`Guild ${guild.id} deleted.`);
     }
 }
-
 module.exports = new GuildDelete();
